Rename the Gateway API client
